Add Empty State View

+++ ToDo/Views/ListView.swift
@@ -12,47 +12,53 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        
+        ZStack {
+            if listViewModel.items.isEmpty {
+                EmptyStateView()
+                    .padding(.top,80)
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List{
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    . onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
+                .navigationTitle("Todo List")
             }
-            . onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-        }
-        
-        .listStyle(.plain)
-        .navigationTitle("Todo List")
-        
+        }    
         .toolbar{
             ToolbarItem(placement: .topBarLeading) {
                 EditButton()
-                    
             }
+
             ToolbarItem(placement: .topBarTrailing) {
                 NavigationLink("Add") {
                     AddView()
                 }
-                
             }
         }
+    //    .tint(.primary)
+
         .alert(item: $listViewModel.alertItem) { alertItem in
             Alert(title: alertItem.title,
                   message: alertItem.message,
                   dismissButton: alertItem.dismissButton)
         }
     }
-        
-        
 }
 
 
 #Preview {
-   NavigationView {
+    NavigationStack {
         ListView()
     }
    .environmentObject(ListViewModel())
